Moved class to workflows
