+++ penelope/notebook/co_occurrence/main_gui.py
@@ -3,10 +3,10 @@ from typing import Callable, Optional, Union
 import ipywidgets as widgets
 from IPython.core.display import display
 from penelope import co_occurrence, pipeline
+from penelope.workflows import ComputeOpts
 from penelope.workflows import co_occurrence as workflow
 
 from .. import co_occurrence as co_occurrence_gui
-from ..interface import ComputeOpts
 from ..utility import CLEAR_OUTPUT
 from ..word_trends.interface import BundleTrendsData
