+++ penelope/scripts/co_occurrence.py
@@ -1,7 +1,6 @@
 from typing import List, Optional, Sequence
 
 import click
-import penelope.notebook.interface as interface
 import penelope.workflows.co_occurrence as workflow
 from loguru import logger
 from penelope.co_occurrence import ContextOpts, to_folder_and_tag
@@ -10,6 +9,7 @@ from penelope.pipeline import CorpusConfig
 from penelope.pipeline.phrases import parse_phrases
 from penelope.scripts.utils import update_arguments_from_options_file
 from penelope.utility import pos_tags_to_str
+from penelope.workflows import interface
 
 # pylint: disable=too-many-arguments, unused-argument
