+++ penelope/workflows/__init__.py
@@ -1,1 +1,2 @@
 # type: ignore
+from .interface import ComputeOpts
