+++ penelope/workflows/co_occurrence/compute.py
@@ -4,7 +4,7 @@ from typing import Optional
 import penelope.co_occurrence as co_occurrence
 import penelope.pipeline as pipeline
 from loguru import logger
-from penelope.notebook import interface
+from penelope.workflows import interface
 
 POS_TAGGED_FRAME_FILENAME_POSTFIX = '_pos_tagged_frame_csv.zip'
