+++ tests/pipeline/spacy/pipelines_test.py
@@ -7,8 +7,8 @@ import penelope.workflows.co_occurrence as workflow
 import pytest
 from penelope import corpus as corpora
 from penelope import pipeline, utility
-from penelope.notebook.interface import ComputeOpts
 from penelope.pipeline.spacy.pipelines import spaCy_co_occurrence_pipeline
+from penelope.workflows.interface import ComputeOpts
 
 # pylint: disable=redefined-outer-name
